requirements.txt
stable-baselines3>=2.0
gymnasium>=0.29
pandas>=1.5
numpy>=1.23
matplotlib>=3.5
PyYAML>=6.0
tqdm>=4.60
torch>=2.0
